WordWrapper - Trying a recursive approach for 2nd version
Also, in the previous I considered the object to be a string wrapping
manipulator extension where you have a string and can manipulate it
multiple times.

In this I went the other way around.
I have a wrapping manipulator where the object is a specific wrapper
(column-wise) and can manipulate incoming strings.
Actually, recursion would have to be this way since the wrap column is the
static thing and the string acted upon.