PerfectNumbers - Calculate set of factors of a number
Trying to use a static method.
Want to see how it feels.
Could be method in original/solution class.
But this could be method to be shared.
Felt trivial to add as injectable dependency.
I believe static methods should be always second guessed but also
not against them.
Alternatively, this could be an inner object that is  constructed as needed.
Semantically the same, is implementation detail that needs not be injected,
can be reused by others and also doesn't bottleneck calls to it if that
is an actual issue in today's world of programming languages.